+++ src/backend/PeopleHub.Application/Services/IFeedService.cs
@@ -4,5 +4,5 @@ namespace PeopleHub.Application.Services;
 
 public interface IFeedService
 {
-    Task<IReadOnlyCollection<FeedPost>> GetFeedAsync(int userId, CancellationToken cancellationToken = default);
+    Task<IReadOnlyCollection<FeedPost>> GetFeedAsync(long userId, CancellationToken cancellationToken = default);
 }
