+++ src/backend/PeopleHub.Domain/Repositories/IFriendRequestRepository.cs
@@ -4,13 +4,13 @@ namespace PeopleHub.Domain.Repositories;
 
 public interface IFriendRequestRepository
 {
-    Task<FriendRequest> GetAsync(int id, CancellationToken cancellationToken = default);
+    Task<FriendRequest> GetAsync(long id, CancellationToken cancellationToken = default);
 
-    Task<FriendRequest> FindBetweenAsync(int userId, int otherUserId, CancellationToken cancellationToken = default);
+    Task<FriendRequest> FindBetweenAsync(long userId, long otherUserId, CancellationToken cancellationToken = default);
 
     Task AddAsync(FriendRequest request, CancellationToken cancellationToken = default);
 
     Task SaveStatusAsync(FriendRequest request, CancellationToken cancellationToken = default);
 
-    Task DeleteBetweenAsync(int userId, int otherUserId, CancellationToken cancellationToken = default);
+    Task DeleteBetweenAsync(long userId, long otherUserId, CancellationToken cancellationToken = default);
 }
